Wrap more of afw image and table
Wrap enough to eanble the following tests:
- testApCorrMap.py
- testBox.py
- testExposure.py
- testExposureTable.py
- testPolygon.py
- testTableIO.py
- testVisitInfo.py

WARNING: CoaddInputs is not properly wrapped: attempting to access
the "visits" or "ccds" fields will cause a segfault. I will create
a new ticket to address this issue; meanwhile one test
of testExposureTable.py is skipped for now to dodge it.

- Wrap afw::geom::polygon::Polygon
- Wrap more of Calib
- Wrap CoaddInputs but see warning above
- Wrap VisitInfo, including giving the constructor named arguments with defaults;
  this eliminates the need for makeVisitInfo, so reimplement that to raise
  a deprecation warning and then call the VisitInfo constructor
- Wrap a bit more of Wcs
- Wrap a lot more of Slots
- Wrap table::Exposure
- Wrap a bit more of Schema and Source
- Modify imageLib.py to import lsst.afw.table.io,
  thus allowing wrappers for image object that can be saved to tables
  to inherit from table::Persistable and support FITS I/O.

+++ tests/testTableIO.py
@@ -23,7 +23,9 @@
 from __future__ import absolute_import, division, print_function
 import unittest
 
+import numpy as np
 import astropy.io.fits
+
 import lsst.utils.tests
 import lsst.afw.geom
 import lsst.afw.table
@@ -52,8 +54,8 @@ class TableIoTestCase(lsst.utils.tests.TestCase):
         Per DM-8211.
         """
         schema = lsst.afw.table.Schema()
-        aa = schema.addField("a", type=int, doc="a")
-        bb = schema.addField("b", type=float, doc="b")
+        aa = schema.addField("a", type=np.int64, doc="a")
+        bb = schema.addField("b", type=np.float64, doc="b")
         schema.getAliasMap().set("c", "a")
         schema.getAliasMap().set("d", "b")
         cat = lsst.afw.table.BaseCatalog(schema)
